+++ src/commands/node/handlers.ts
@@ -7,7 +7,8 @@ import * as constants from '../../core/constants.js';
 import {type LockManager} from '../../core/lock/lock-manager.js';
 import {SoloErrors} from '../../core/errors/solo-errors.js';
 import {type Lock} from '../../core/lock/lock.js';
-import {LeaseWrapper, type NodeCommandTasks} from './tasks.js';
+import {type LeaseWrapper} from './lease-wrapper.js';
+import {type NodeCommandTasks} from './tasks.js';
 import {NodeSubcommandType} from '../../core/enumerations.js';
 import {NodeHelper} from './helper.js';
 import {AnyListrContext, type ArgvStruct, type NodeAlias, type NodeAliases} from '../../types/aliases.js';
