+++ src/commands/node/lease-wrapper.ts
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: Apache-2.0
+
+import {type Lock} from '../../core/lock/lock.js';
+
+export interface LeaseWrapper {
+  lease: Lock;
+}
