+++ src/commands/one-shot/one-shot-command.ts
