+++ src/core/lock/interval-lock.ts
@@ -6,7 +6,8 @@ import {LockHolder} from './lock-holder.js';
 import {DEFAULT_LEASE_DURATION, DEFAULT_SOLO_NAMESPACE_LABELS} from '../constants.js';
 import {sleep} from '../helpers.js';
 import {Duration} from '../time/duration.js';
-import {type Lock, type LockRenewalService} from './lock.js';
+import {type Lock} from './lock.js';
+import {type LockRenewalService} from './lock-renewal-service.js';
 import {StatusCodes} from 'http-status-codes';
 import {type NamespaceName} from '../../types/namespace/namespace-name.js';
 import {type Lease} from '../../integration/kube/resources/lease/lease.js';
