+++ src/core/resolve-gossip-fqdn-restricted-options.ts
@@ -1,0 +1,13 @@
+// SPDX-License-Identifier: Apache-2.0
+
+import {type K8} from '../integration/kube/k8.js';
+import {type NamespaceName} from '../types/namespace/namespace-name.js';
+
+export interface ResolveGossipFqdnRestrictedOptions {
+  k8?: K8;
+  namespace?: NamespaceName;
+  stagingDir?: string;
+  cacheDir?: string;
+  resourcesDir?: string;
+  applicationPropertiesPath?: string;
+}
