+++ src/data/schema/model/remote/interfaces/component-ids-structure.ts
