+++ src/types/command-flags.ts
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: Apache-2.0
+
+import {type CommandFlag} from './command-flag.js';
+
+export interface CommandFlags {
+  required: CommandFlag[];
+  optional: CommandFlag[];
+}
